src/BE/requirements.txt
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.0
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.2.0
fastapi==0.105.0
Flask==3.0.0
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mlflow==2.9.2
numpy==1.26.2
oauthlib==3.2.2
optuna==3.5.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pyscopg2==66.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
scikit-learn==1.2.1
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
streamlit==1.29.0
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
watchdog==3.0.0
websocket-client==1.7.0
Werkzeug==3.0.1
zipp==3.17.0
